+++ src/main/java/com/github/wasiqb/coteafs/datasource/annotation/DataFile.java
@@ -24,9 +24,8 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Annotation to mark Pojo as data file. It will search the folder for file
- * with name class name (each word in name separated with hyphen (-).
- * Also the file extension must be any of the following:
+ * Annotation to mark Pojo as data file. It will search the folder for file with name class name (each word in name
+ * separated with hyphen (-). Also the file extension must be any of the following:
  * <p>1. yml</p>
  * <p>2. yaml</p>
  * <p>3. json</p>
@@ -37,16 +36,15 @@ import java.lang.annotation.Target;
 @Retention (RetentionPolicy.RUNTIME)
 public @interface DataFile {
     /**
-     * Data file name, if not provided, it will be data class name in
-     * lower case hyphen separated words.
+     * Data file name, if not provided, it will be data class name in lower case hyphen separated words. E.g. If class
+     * name is {@code LoginData} then file name is considered as {@code login-data}.
      *
      * @return Data file name
      */
     String fileName () default EMPTY;
 
     /**
-     * Relative folder path containing data files. If not provided,
-     * it will consider {@code src/test/resources}
+     * Relative folder path containing data files. If not provided, it will consider {@code src/test/resources}
      *
      * @return Folder path
      */
