+++ src/main/java/com/github/wasiqb/coteafs/datasource/error/OperationNotSupportedError.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2020 Wasiq Bhamla
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.github.wasiqb.coteafs.datasource.error;
+
+import java.io.Serial;
+
+/**
+ * Custom exception for coteafs-datasource project.
+ *
+ * @author Wasiq Bhamla
+ * @since 15-Dec-2020
+ */
+public class OperationNotSupportedError extends Error {
+    @Serial
+    private static final long serialVersionUID = -5902785674999190917L;
+
+    public OperationNotSupportedError () {
+        super ();
+    }
+
+    public OperationNotSupportedError (final String message) {
+        super (message);
+    }
+
+    public OperationNotSupportedError (final String message, final Throwable cause) {
+        super (message, cause);
+    }
+}
