--- testng.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright (c) 2020 Wasiq Bhamla
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite name="Data Source Test Suite">
-    <test name="Test">
-        <classes>
-            <class name="com.github.wasiqb.coteafs.datasource.DataSourcePropertiesTest"/>
-            <class name="com.github.wasiqb.coteafs.datasource.DataSourceXmlTest"/>
-            <class name="com.github.wasiqb.coteafs.datasource.DataSourceYmlTest"/>
-            <class name="com.github.wasiqb.coteafs.datasource.DataSourceJsonTest"/>
-        </classes>
-    </test> <!-- Test -->
-</suite> <!-- Suite -->
